+++ CHANGELOG.md
@@ -16,6 +16,9 @@ Increment the:
 ## [Unreleased]
 
 * [BUILD] Upgrade opentelemetry-proto to v0.19.0 [#1579](https://github.com/open-telemetry/opentelemetry-cpp/pull/1579)
+* [METRICS EXPORTER] Add `OtlpGrpcMetricExporterFactory` and `OtlpHttpMetricExporterFactory`.
+  [#1606](https://github.com/open-telemetry/opentelemetry-cpp/pull/1606)
+* [METRICS EXPORTER] Add `OtlpGrpcClient` [#1606](https://github.com/open-telemetry/opentelemetry-cpp/pull/1606)
 * [BUILD] Fix header only api singletons [#1604](https://github.com/open-telemetry/opentelemetry-cpp/pull/1604)
 
 ## [1.6.0] 2022-08-15